Generating correlated random variables with Cholesky decomp.
This version is generating gaussian random variables that follow the
covariance matrix used as input. I checked that the mean is zero as it
was supposed to be, the covariances match the covariance matrix and that
the distribution is gaussian (it looks gaussian and a kolmogorov-smirnov
test returns high values).