Start from Cov. matrix of lognormal variables
This version takes as input a covariance matrix of lognormal variables,
compute the covariance matrix of the associated gaussian variables and
use the Cholesky decomposition to generate the gaussian variables that
follow the gaussian covariance. As input we are taking two new vectors:
the mean and the shift of each variable. The conversion from lognormal
to gaussian covariance has benn tested for a simple case (3x3) where
all means and shifts are the same.

Next step is to transform the gaussian variables into lognormal variables.