Fixed wrong expsigma2 calculation
Tests made with different entries for means and shifts revealed that
the term expsigma2 used to calculate the lognormal variables from gaussian
ones was wrong. It was giving the correct answer for mean=0 and shift=1 by
accident. The formula has been corrected, and it seems to be working well.
The same tests as the ones used for the previous version were made, this
time with mean values {0, 1, 2} and shift values {1, 2, 0}. All worked.
Also added the option for outputing the gaussian covariance matrix.