Creation of gencovl
This version includes the gencovl routine that takes several C(l)s
between different fields and redshifts and creates covariance matrices
for each l, since for a full sky these are independent. The basic
algorithm seems to be working fine: I compared the generated covariance
matrices with the input C(l)s for a few cases, and they were making sense.

Next step is to better organize the routine, include a configuration file
and some clever form of bookeeping for which C(l) went to each element of
the output covariance matrices.

A few changes were made to other files that might affect the behavior of
the corrlnfields, specially the input file. This should be checked as well.