gencovl Completed
In this version we added a configuration file scheme to gencovl. Both
gencovl and corrlnfields use the same config file, which might be good
or bad, let´s see. We also included an output of the order of the fields
as variables in the covariance matrix. This version seems completed
up to future changes when those are needed.

Next step is to work on the corrlnfields routine.