Implementing transformation for lognormal Cl to gaussian Cl
This version is in the way of implementing the transformation of
the Cls from lognormal variables to the associated gaussian variables.
This is done using the Discrete Legendre Transform (DLT) by s2kit 1.0.
The code has been tested in the directory s2kit that is not being
tracked. We sucessfully calculated the correlation function
(by comparing to the mathematica output) and we were able to recover
the Cls after two transformations. Tests were ran in the hx-test.cpp
(not tracked), and the necessary functions from s2kit were packed in
the file s2kit10_naive.cpp (we are using the naive implementation of
the DLT).

We are also migrating the gencovl code to corrlnfields so no intermediary
files must be written.

NEXT STEP: Implement the local transformation of the lognormal correlation
function to the gaussian one. For that we will need the means and shifts
of the fields. It seems that there is a better way of organizing these
informations, possibly by packing everything in one single file, although
this might cause complications if one wants to generate only gaussian
catalogs that do not require shifts. Things to remember:
- Instead of applying the transformation mentioned above in the covariance
matrices for each ell, we are applying to the correlation functions. It
should be the same.
- Before copying the symmetric parts of the covariance matrix, we are computing
the gaussian Cls in order to save computing time. Later these must be copied.