Transformed C(l) lognormal to gaussian
In this version the C(l)s for the auxiliary gaussian variables is
calculated. It is still needed to check if the transformation was done
correctly. I compared the intermediate steps of obtaining the correlation
function of lognormal and gaussian variables. Last step (the computation
of C(l)s for the gaussian variables) needs checking. It would also
be good to double-check everything since a few small changes were made in
the array sizes (and therefore bandwidth of DLT).

Next steps:
- Test like said above;
- Organize data into covariance matrices;
- Check if previous TODO list was completed.