Changed cov matrix and friends offset from 1 to 0
In previous versions, the matrix and vectors storing elements of
the covariance matrix started at index 1, which looks nicer for me.
Unfortunately, the standard is to start at 0, and gsl used this
convention. To avoid confusion in the future between different offsets,
I changed the starting index to 0. Ran the code to test for bugs and
compared its output with previous ones. Everything is fine.