Corrected means (I think), but recovered C(l)s are slightly off
This commit was done after Christmas and New Years but the code was
written before that, so I don't remember exactly what is going on.
I believe the issue with the means and shifts mentioned in the previous
commit has been corrected. The present code generates a lognormal map and
compute the alm's from it. The recovered alms are a little off from the
original one, and this is being investigated. To investigate we will generate
many realizations of the fields to decrease the significance of statistical
fluctuations. To do that, we need a code that can be paralelized in a stupid
way (running many processes). TO do that we need to change how the program
lists the input files (it cannot output to a file as it is currently doing).