Removed use of data written to HD and other stuff
In this version the code does not need to transfer data using files
written to the HD. This enables us to parallelize the code by simple
dumb calling many times the process with different data. Specifically
to analyse the recovery of alm's, we created shell and python scripts
that run and compute alm's many times over. We also added to possibility
to export the auxiliary catalog to a file.