Magically recovering C(l)s the correct way
After much fight against the code, it decided to be more friendly and
output the correct lognormal C(l)s with the correct formulas behind it.
No idea of why it was not working before. Some embarassing pieces of
code used to debug are still lingering around and must be removed slowly.
I have not retested the lognormal properties of the field after this
debuging but it should be working fine.

TODO:
- Remove debuging leftovers without compromising the results.
- The testing of the results must be simultaneously for the lognormal
distribution of the generated fields and the recovered C(l)s.