Mean projected galaxy density implemented
I implemented a function that computes the average projected galaxy density.
It uses pieces of code previously written: the Cosmology object which is
almost a simple struct full of cosmological parameters; and functions that
calculate the comoving distance. The projected density currently assumes that
the angular contrast density is for galaxies (whereas the contrast density is
actually the metter contrast density), that is, it ignores HODs or biases.
In other words, it assumes a constant multiplicative bias of 1. Besides, the
redshift selection function is currently constant inside the integration interval
(i.e. top-hat) and the formulas should be double-checked.

TODO:
- Implement a more realistic relation between matter contrast density and
  galaxy contrast density;
- Verify the intervals and weighting functions for projecting the mean
  galaxy density;
- Check if the formulas used to compute the mean projected density is correct.