Angular sampling inside pixels visually tested and working.
I sampled uniformly the angular positions inside various pixels
using the RandAngInPix function (or some name like that) and made
plots of the original pixel and of a finer map with the sampling.
The angular sampling looks fine, without artificial clustering on
the poles, pixel borders or sampling outside the pixel.