Generating catalog
This version generates a catalog of galaxies that has been Poisson sampled
from the density field and assigned a random redshift and angular position
inside the pixel. The angular position is fine and the redshift must be
improved (currently it is uniformly sampled). To each galaxy we also assign
the convergence of the pixel. The catalog seems to be generated corrently but
more tests are welcomed.