Trying out shear computation by convolving convergence map
This version includes a test of calculating the shear map from the
convergence map by convolving the map around each pixel with a
quadrupole. It is not working properly since it does not converge
to a fixed value with an increase in the radius of truncation. We will
try a different approach in the following version, which transforms
the convergence map back to harmonic space, use its coefficients to
compute the shear map harmonic coefficients and transforms back to real
space. This approach is based on Wayne Hu's PRD Vol. 62 p. 043007 (2000).