GeneralOutput implemented
In this version I replaced almost all output to files for functions coded
in the GeneralOutput.cpp file. This is meant to clean the main code and to
ease the future output of similar things. The output have been tested
(compared with previous ones) and they agree. The fits output might disagree
but it seems to be due to a difference in the date of file creation.