Catalog generating version
This version generates a catalog with galaxy position, redshift and
ellipticity which is obtained by combining intrinsic ellipticity with
shear. The intrinsic ellipticity is drawn from a gaussian. The previous
bug was removed. The shear distribution seems slightly odd (the Q mean
sometimes differ from zero and there are apparent modulations of the
shear orientation), this needs to be further checked. We also removed a
few unused variables and found out an error on deallocating the gaussian
variables gaus0 and gaus1. Next step is to test the code with more l's
and more fields. Then we could implement bias and other stuff.