Regularizing cov. matrices
We finished the implementation of a new selection function, and as
far as I remember everything is working fine. There are issues when
building the catalogs, but we are not addressing this now since we
are focused on regularizing the auxiliary gaussian covariance matrices.
We have implemented two methods, the minimization of Frobenius norm
(which translates to setting negative eigenvalues to zero) and a
downhill minimization process through steps. The first was was tested for
one matrix and compared to Mathematica output, and is working fine.
The second one seems to have problems, as its results differ from
Mathematica's for certain matrices.

TODO: correct bug that causes such problems.