Regularization corrected
The Regularization method of successive approximations was fixed: the
problem was that the eigenvalues were not sorted according to their
absolute value. Now they are sorted. The fact why this is necessary is
still a little mysterious and it doesn't really make sense since it
is not possible to assign an absolute order for the eigenvalues in a
way that eigenvalues for different matrices can be associated one
with another in a meaninful way.