Catalog building for many z bins fixed
In previous versions the shear maps for various redshift bins were not
following the redshift of the galaxies. Now this was fixed, so galaxies at
a redshift bin i receive ellipticities according to the convergence and shear
maps in that bin.

TODO: The redshift selection function in the separable case still needs to
      be loaded in a more clever way to save memory, even though there should
      be no errors in the current implementation, just memory inneficiency.

We also intend to implement the computation of a lognormal Cl based on the
regularized auxiliary gaussian ones.