Regularized input lognormal Cls
In this version the take the regularized covariance matrices for the
auxiliary gaussian variables and reverse the transformations applied,
thus reobtaining the input Cls that would lead to a decent auxiliary
covariance matrix. We tested that the inverse transformation is
working fine by doing it without regularizing the covariance matrices.
The inverse transformation returns very good fractional error except when
the Cls get to zero (where the use of fractional error is a bad choice).
Therefore we conclude that the regularized input Cls are good representatives
of Cls that would lead to well-behaved covariance matrices.