Stability changes
In this version the selection function includes member functions to check its Nside and
pixel ordering Scheme. Moreover, the regularization process of the input Cls allows for
no-use of the exponential suppression. The python routines for organizing alm output
and computing the recovered Cls was generalized for an arbitrary number of fields.