Testing version for lognormal limits
This version was modified so the allowed limits on the correlations of
lognormal fields could be tested. The tests were ran in the onlykappa
directory. They are performed through the editing and running of the
varlog.sh script. It basically scales the input cross Cl (just two fields)
and changes the shift parameter and extracts from the corrlnfields code
the eigenvalues for the gaussian covariance matrices.