Freed most of the code from fnz matrix
Instead of recording in a table which field and redshift is in which position
and using it to recover the information, we modified most of the code to
use the n2fz function. Now the fnz matrix is only used in the beginning of
the code to make sure the Cls are being read right, without lack of Cls and/or
excess of information.