Cl processing modularized
In this version the input Cls loading, organizing, transforming for
gaussian variables, etc., is all inside a function called ClProcess.
This is meant to make it easier to skip this processing in case
we are loading pre-computed Cholesky triangular matrices. I will test
this code by comparing the output with those of previous versions.