Working loading mixing matrices version
This version is able to load mixing matrices already created from
Cholesky decomposition and use these to generate alm's. A few tweaks
on GSLMatrix output precision were made to ensure the results are the
same as doing the Cholesky on the fly. Moreover, a few screen dump
formatting was done and a few warnings introduced regarding output.