Few tweaks and corrections after parallelization of DLT weights
This version includes a few corrections (like that if the regularized
lognormal Cls are requested, then the regularization must be performed
in all ls and not just on the ones used in alm generation) and tweaks
(like organization of output information on screen). Tests were run
to make sure that the auxiliary gaussian alms are the same as the serial
ones. We also increased the precision for the mixing matrices output
in order to ensure that the aux-alms generated from loaded mixing matrices
and calculated on the fly ones are the same.