alm generation parallelized
This version the auxiliary gaussian alm's are generated in parallel.
Due to the limitation that the use of random number generator must be
done sequentially, this did not improve the performance. Better to
move back to serial on this part.

PS: the alm's were compared to the serial ones and they are the same.