Regularization maximum iterations limit
We introduced a limit to the maximum number of iterations that the
stepper regularization method can do. If the limit is crossed, a
warning is displayed. The prepClassInput.py was changed as well:
the factor for transforming from lensing potential to convergence
gained a -1 sign, as it is expected from theory (I think). However,
the output is weird. This has to be inverstigated still.