Load 1 radial selection function per galaxy type
I improved the Selection function class code so it now only loads
one radial selection function (for the separable case) per galaxy type
(before it was loading one -- the same -- for every redsfhift).
I tested for memory leackage (since an extra dynamic array had to be
introduced in the class) and it worked fine.