Auxiliary alms generation re-parallelized
I found a way to parallelize the generation of alm's by using different
number generators for each thread. They are constructed in a way to avoid
correlations between runs with different seeds and to be reprodutible for
the same seed (if the number of threads is the same). We tested the average
properties of the generated samples, they seemed fine, and the reproducibility,
also working.