Parallelized Poisson sampling
The poisson sampling of the galaxy fields are now parallel, implementing
the same ideas from the aux. alm's generation. We checked for reproducibility
and overall aspects of the final catalog. Speed was improved. We also put
random variables memory allocation outside the aux. alm loop, to improve
speed. Not much was gained.