Working redshift sampling according to selection function
In this version the random sampling of redshifts in the bin was improved
from uniform to following the selection function.

NOTE: Step functions in the selection function may lead to wrong number
of galaxies in the bin if the break does not coincide with the bin limits.
This happens because we use the bin central value to represent its value
during the map generation and not the mean (or something like that).