Star Mask, selection function scaling and just booking selection functions implemented
In this version we added the possibility of multiplying the selection
function by another selection function (fixed for every z) that may represent
a star mask. We also added the possibility of rescaling the selection
function by a constant factor to ease increase or decrease in number density of
galaxies. We also introduced the possibility of loading the selection
function but not apply to the catalog the angular part (including the
Star Mask), only the radial part. For non-separable selection functions,
this option means that the star mask is ignored. Even if this seems like
a good option, the use of 'just bookkeeping with non-separable selection
functions is currently blocked in the SelectionFunction.load by an error.