Implemented option that outputs maskbit to catalogue
If the kerword 'maskbit' is inserted in the CATALOG_COLS list,
the catalog gains a column with a integer that tells if the
galaxy falls into the star mask, selection mask, both or if
the selection or star mask has a value different than 0 or 1.
Tested, seems to work. We also changed the PrintTable formatting.