Skip recalculation of klm's for gaussian realizations
I remembered that the klm's of the convergence field must be obtained
from the map only for the lognormal realizations because of the
exponentiation of the pixels. For gaussian realizations there is no
operation on the pixels apart from adding the mean which only affects
the monopole which is irrelevant for shear E and B modes, so we don't
need to go to real space, the transformation from klm to E modes
can use the results from generating the auxiliary gaussian alm's
(which, for the gaussian case, are the real thing and not auxiliary).
Thus, I optimized the code for gaussian realizations by skipping
the reobtaining process for the klms.