Tweaks for compiling on Splinter
We removed collapse options on pragmas and made the nested for loops
a single one to improve the portability of the code for older
OpenMP versions. We also introduced an output for the first random
numbers of the rnd[0] (the serial seed) and #defined the maximum input
seed. We also made some updates on the scripts for making many realizations
of recovered alms.