Parallel catalog working (tested briefly) & new process timing
We corrected the reproducibility of the parallel catalog generation
by privatizing the processor index l. We checked that two runs
were giving the same catalogs. for things that did not depend on random
numbers such as pixel number and kappa, we compared before and after
paralellizing and it matches.

We also changed the timing of processes because the clock function
goes crazy with parallel processing.