Parallelized exponentiation and addition of mean to pixels
We parallelized the loop over the pixels (but not the outer loop over
the fields) when exponentiating (LOGNORMAL) and adding the mean
(GAUSSIAN). We tested for LOGNORMAL case and there was a small change in
only one entry, maybe due to the implementation of the parallel sum
through the reduction clause. Everything seems fine. We did not test
the GAUSSIAN case, but it is pretty simple and should work.