Created Xi calculator
In this version we created a program in C++ that computes the correlation
function from a catalog. This preliminary version has several limitations:
it necessarily assumes that the columns are DEC, RA, ..., FIELD VALUE, ...
where DEC and RA are given in degrees; it only computes the correlation
between values of the field which is assumed to have zero mean. It cannot
deal with non-zero means and with galaxy position correlation. On the
positive side, it is a parallel code.