Cl recovery and poissoned MAP output
This version outputs the maps after poisson sampling if they
are requested (NOT TESTED) and computes and returns the recovered
Cl's from the maps without noise (TESTED, working). There is a
deviation between the recovered Cl's and the original ones for
lognormal realizations (and not for gaussian ones). Moreover, the
power spectrum for the gaussian case is recovered without any
multiplication with the pixel window function, which is weird.