Auxiliary covariance matrix testing (including signal change)
This version includes test of properties of the auxiliary covariance matrix:
positive diagonals and correlations bounded by -1 and 1. Moreover, we
checked if the matrix elements change signals when going from lognormal to
auxiliary gaussian matrices, thinking that this might cause problems.
We verified that the signs can change and not cause problems to the cholesky
decomposition. This means that this test is unecessary. Maybe this makes sense
since the Cls are mixed in the process (going from lognormal to gaussian is
a non-local transformation in harmonic space), so going from correlated to
anti-correlated alms is not that weird.