Catalogue # of galaxies from INT to LONG
The process of counting galaxies and creating a catalogue now uses
long variables instead of int since the number of galaxies might surpass
2.1 billion, which is near INT_MAX. We tested that the current version
returns the same catalogue as the previous one for ~1 million galaxies.
We also introduced a warning if NSIDE is greater than ~13000, which would
lead to Npixels>INT_MAX (Npixels remained int).