Changeable precision (FLOAT or DOUBLE) for alms, maps, and catalog
Now it is possible to change the precision of Healpix alms and maps
independently and the precision of the galaxy catalog by changing
the definition of ALM_PRECISION, MAP_PRECISION, SEL_PRECISION and
CAT_PRECISION in definitions.hpp file. We tested that maps generated
now and before are the same when using doubles, and they are almost the
same (checked a few entries) when changing from double to float.
We did not test other outputs, only ASCII maps. The code indeed used a lot
less memory (~half) and appeared to be faster, although not a lot or at all.
We also changed the catSet matrix from int to char to save memory,
promoted PrintVecs from double to template, and moved global definitions
from corrlnfields_aux to definitions.hpp.