Fits catalog output tweaks and skip add 0 mean in gaussian
We introduced a few corrections in the FITS catalogue output regarding
RA and DEC and the FLOAT precision for the catalogue. The solution demands
setting an extra definition in definitions.hpp, so its not the best option,
but it works. We also included an IF in the "add mean" loop for the gaussian
realisations to skip in case of zero mean. We did not test it but the change
is simple and it should work.