Tiny term to cov. diagonals
We introduced the possibility of adding a term to the covariance matrices
diagonal in case the Cholesky decomposition might fail. This was needed because
of numerical approximations that make the Cholesky decomposition fail even
when the eigenvalues estimation routine returns positive ones. A few other
tweaks were made also, I think.