Output for integrated density alm's and Cl's, Healpix map2alm weights.
In this version we included the possibility of using Healpix weights saved
in fits files in Healpix data directory instead of ones when doing the
map2alm transform. We still need to check that we are using the weights
correctly (ask Boris and test the lognormal shear computation; we verified
that the use of Healpix weights leads to very small effects on density and
convergence recovered alm's and cl's).

We also generalized the computation and output of recovered alm's and cl's
to map arrays that contain empty maps (which is the data structure used for
integrating density along the LoS). All this was put inside a function RecoverAlmsCls
so we could use the same process for recovering both usual fields and the
integrated density. The output was verified to work as previously as well as
in the case of empty maps (integrated fields), and one realization of the
integrated density matches with the convergence input Cl's very well.