Kappa from integrated density 2.0: get it for sources at all z bin ends.
We generalized the previous code for computing the convergence from the
density by LoS integration to zsource at the end of each redshift bin and
not just the last one. We ran a test by comparing the map, alm and cls
output to previous ones, making sure the results for the last redshift bin
is the same and that the number of columns is the expected one (the test
was performed with 2 bins -- proofconcept directory).

We also acknowledge the removal of many files in data directory and others
that should not be tracked.