Implemented but turned off stat limitation for gaussians
The last commit mentioned that statistics for auxiliary gaussian fields
should not be print for gaussian fields. We introduced a way of doing
that in the code but concluded that that argument does not make sense as
you can always fit a gaussian field with a lognormal. So we turned this off.