kappa-delta X-Cl changed sign, my own shift formula
From the ini2info.py file, we removed the need for a table of
values that was missing in Splinter. Moreover, I added a new formula
for the shifts based on the sum of lognormal variables, and this
formula is now the default. Finally, I realised that the cross spectra
between convergence and density had the wrong sign which was introduced
in the prepClassInput.py for transforming CLASS lensing potentials to
convergence. Numerical integration of density auto-Cls weighted by
weak lensing kernel returns the cross-Cls and this can be used to check
for consistency. This was done and we confirmed the previous sign was wrong.
Now it is correct.