regCl acts only on N(N+1)/2 Cls (avoid repetition)
To avoid spitting identical files [C_l^{ij}=C_l^{ji}] and wasting computing time
we changed the loop over the recovering Cls to avoid repetitions. We compared
previous and latter output to make sure they are OK, and they are.

We also checked that the fractional change output on screen is actual fractional change
and not percentage change. We corrected the text to avoid confusion.