Created Fields database structure
The class FZdatabase was introduced to enable the code to deal with
an arbitrary number of redshifts for each field (that might be different)
and with arrangements of numbering that do not start on 1 and are not
necessarily contiguous. We tested the code with the test.cpp file \
(which should be deleted later) and it is working fine.