FZsdatabase implementation up to selection function compared with previous and working
We compared all previous output that does not depend on the selection function
to the current output that uses FZdatabase. They all agree (small corrections were made
to the intdens part). FITS files disagree when using diff maybe because of date stored
somewhere because the maps themselves do not differ at all when using ViewMap.py.