Shear are computed before Poisson and selection function
We changed the order of the computations: the shear is computed from convergence
before the use of selection function and poisson sampling. This makes it
independent of the selection function. We also introduced a function that tells
if computing shears is necessary and use it to condition the shear calculations.
We compared catalogs generated before and after this inversion of order and they
are the same. We did not test if the condition for computing shear is working
properly as it should be easy to spot problems in the future.