Changing catalog creating process
In this version the catalog creation process writes the info about all galaxies
in each cell to a matrix buffer and then writes to the file. It does not
create a full catalogue in memory before writing it to a file. This change should
save memory (since only the buffer is saved instead of the whole catalog) and
may be faster since we avoid counting the total number of galaxies a priori.
However, the process of making the catalog is not parallelized here.
We have not compared the results yet to check that it is working fine and
have not implemented the output to fits file.