Fix catalog minor output bugs and using map2alm_iter 1 for shears
When l>2*Nside or something like that, the shears in the lognormal case
were getting an abruptly larger variance in the poles (|dec|>45deg).
This happens for map2alm_iter with iter=even and does not happen with
iter=odd. We are now using iter=1 for getting alms from lognormal kappa,
used to compute shears.

We also corrected a few bugs in the catalog output: galtype was wrong since
the use of fieldlist class, now it is working; and fits output for kappa was
getting overflow, now corrected.