Dens2KappaCls and organization
The main addition in this version is the Dens2KappaCls code, that
allows to compute the expected kappa-kappa Cls and kappa-density
Cls that would be obtained when kappa is determined as a riemann
sum of densities; the expected value is obtained from the density
auto and cross Cls. We compared the output of Dens2KappaCls with the
recovered Cls from corrlnfields code and they agree apart from a
~0.4% difference at high l (the Dens2KappaCls result is below the
recovered results using LoS integration).

Within this theme, we realized that using the central redshift bin
value for the lensing kernel is a biasing approximation since the
Kernel is concave. We created the function AvgKappaWeightByZ that
uses the average weight in the bin instead. And the results for
Dens2KappaCls are better using this function. In this version,
Dens2KappaCls uses the AvgWeight while the corrlnfields still uses
the central one.

We also reorganized a few function between files and changed the
passing of Cosmology object to functions from pointer to const ref.