Exp. for lognormal fields more efficient, Spline introduced.
Instead of computing the mean and variance of the gaussian field from the maps, we use the
mixing matrices to compute the gaussian Cls and use these Cls to compute the theoretical
value for the Gaussian variance. This is expected to be faster than the former approach.
It also enabled us to combine the fields and pixels loop into a single one which required
long variables. We verified that the output obeys the expected statistics like mean,
distribution shape (qualitatively) and shift parameter. We also compared with previous
realizations and made sure they are similar.

We also began to introduce the Spline for computing the theoretical power spectra
expected from density LoS integration. The function Spline is tested qualitatively and
it is working.

!! However, it would be good to test it more quantitatively.

We introduced a code in Dens2KappaCls to compute the LoS integration convergence
using the spline, but right now it is commented. Dens2KappaCls is still using the
AverageWeights inside the redshift bins.