Added BADCORR_FRAC
The configuration input BADCORR_FRAC is a fraction that is added to the respective
variances in the aux. cov. matrix if they result in a correlation greater than one
or smaller than minus one. This should be small and can affect GAUSSIAN simulations
as well. This is important for neighbouring high redshift convergence Cls in thin bins,
that are highly correlated.