extrapolateCls finished, Dens2KappaCls parallelized and comments put in other scripts
The script extrapolateCls.py was finished, even though the exact form of extrapolating
the results from the lowest z bin to zero is not super well stablished, a lot of subjective
bad approximations: the extrapolation is linear, the auto cls are thought of as dirac
delta functions, and the off-diagonal terms in the covariance matrix by ells are thought of
as planes. There is also the issue that when the bins have different widths, it is not
clear which width should be used.

The Dens2KappaCls was parallelized over the multipoles. It is probably not the best
parallelization scheme, the best probably would be over the sources redshifts (outer loop),
but this is better than previous speed and we have the advantage that the output is serial.