Input Cl Gaussian smoothing
We introduced the possibility of changing the input Cls according to a Gaussian smoothing
function that is applied to the fields; this change to the Cls is usually called Window
function and is the square of the harmonic transform of the Gaussian smoothing function.
The Gaussian width is controled by WINFUNC_SIGMA and the output by SMOOTH_CL_PREFIX.
We tested the behaviour of the code by varying these two options and compared the results
for one fixed sigma to the same method using mathematica.

TODO: It would be interesting to generate a field without the window function,
apply the smoothing, recover the cls and compare with the CL with window function.