Increased config file string sizes and document code caveats
Current implementation of ParameterList uses static allocation for strings
passed to configuration file keywords. Previous maximum size was 70, I increased it
to 110. The 110 limit was added to Flask manual in the configuration file and keywords
section. This can be completely solved by reqriting ParameterList using dynamic
memory allocation, possibly with the std::string class.

It was noticed by Sree that the compilation would fail for Healpix version 3.30. The
issue is that Healpix changed xcomplex in alm class for std::complex when moving
from version 3.20 to 3.30. I updated the manual to mention this limitation.

TODO: Change ParameterList to use std::string instead of char[]
for reading string keywords to avoid crashing due to long strings.