Scripts docstring to stdout, zrange2camb
I introduced the possibility of printing to stdout the docstring of the
python scripts if wrong (or none) number of input parameters are passed
to the scripts. I have tested that they print works for no input, but haven't
checked if I accidentally blocked the funcionality by demanding the wrong
number of parameters.

I also included the script z-range2camb-ini.py that eases creation of
CAMB sources .ini files for counts.