Allow REG_CL_PREFIX output for DIST: GAUSSIAN
In this version the code can output regularized Cls even if the fields
are Gaussian. This allows unsuited input Cls to be corrected for defects
even when simulating Gaussian fields. We would not expect great changes in
the input Cls for Gaussian fields but there might be numerical distortions
to the Cls, therefore this possibility is useful. The manual was changed
accordingly.

Some other tweaks regarding memory allocation and code commenting were
made.