Selection function uses integrated n(z) instead of n(zc)*dz
The code now uses the integral of the radial selection function
in the zmin zmax interval to compute the selection function that will
multiply (1+mapf[i][j]). Before we used a worse approximation of using
the z bin central value times dz.

TODO: allow for both ways to be used according to choice. This is better
implemented inside the selection function. Bring dz and dw inside.
TODO: change Manual to explain all this.