Allow compatibility with Healpix version 3.30 and newer
I introduced precompiler #if based on Healpix version to use xcomplex.h
old Healpix functions or std::complex. I haven't tested it yet nor updated
the manual. This should be done right away.

I also changed the way the git version is written to the code by
using the -D g++ option. This is cleaner and it works. I still need
to register if the code was modified from the commit.