Gen. alm bug fixed, updated example.config and manual
Finished implementation of comatibility for healpix versions 3.30 and
higher by adding the info to the manual and fixing a bug from previous
version (which was giving completely wrong results). This version has
been compared with results from previous one and they match perfectly
if the Healpix version is not changed. When moving from 3.20 to 3.31,
the results change a bit but always below a decent amount of precision.

I have not tested other binaries aparto from flask, but in principle
they should continue to work.