Condensed catalog angular coords. options to function and allow for distance r in catalog
To better organize the code, we placed the options regarding RA,DEC,THETA,PHI in the
catalogues inside a function OrganizeAngularCoord. The code was tested for a few options
and it seems to be working fine.

We also allow now for the catalog to contain a column called 'r' which is the comoving
radial distance to the galaxy according to its specific redshift and the cosmological
parameters OMEGA_m, OMEGA_L and Wde provided in the config file. The distance is given
in h^-1Mpc. Note that if the input Cls include RSD, this distance will be in redshift space,
I guess. I tested the behaviour with and without r in CATALOG_COLS and everything was smooth,
and I compared the distances output with Mathematica's formula, and it agrees.

TODO: Update the manual and example.config, including reference about the cosmological
      parameters and about RSD.