Bug fixed on undersampled SELEC_Z_PREFIX
If the spacing between tabulated points in the radial selection function
SELEC_Z_PREFIX was equal to or larger than the redshift bins being simulated,
the integral os the radial selection function would give a wrong value. We
have corrected this bug and checked that even when the radial function is
undersampled, the integral now gives the right answer.